requirements.txt
moviepy==1.0.3
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
